consider an individual as favorite if one of its persona is a favorite
We used to consider all the personas as having the "is-favourite" property
defined. So _update_single_valued_property was giving the priority to the
primary store (usually EDS) leading to the individual being considered as
unfavourite even if it has a Telepathy persona as favourite.

We now consider the "is-favourite" property being defined if it's set to TRUE
so any favourite persona will take presedence.

https://bugzilla.gnome.org/show_bug.cgi?id=705196