Use autotools's parallel test harness
We were already unintentionally using it in some cases (depending upon the
developer's system) without the proper safeguards in place. Without them,
the EDS and Tracker tests will mysteriously result in a ton of segfaulting
background processes because they can't be safely run simultaneously.

The upside of using the parallel harness is that it's still safe to use for
the simpler test suites (eg, key-file, folks, telepathy) and should speed up
their runtimes quite a bit. It also adds cleaner logging facilities and
makes the default output a little cleaner.

Bug: https://bugzilla.gnome.org/show_bug.cgi?id=679826