ofono: Work around a Vala bug causing a double-free when adding a modem
Work around Vala bug #709665:
	https://bugzilla.gnome.org/show_bug.cgi?id=709665
which allowed an assignment of a string*[] value to a string[] variable.
string*[] values need the container freeing, but not the members of the
array. However, since the variable was a string[], Vala generated C code to
free both the container and the members, resulting in a double-free of each
member.