core: Tighten up definition of Persona.display_id to ensure it’s unique
It is defined as being unique, so cannot be a human-readable full name,
since those can be non-unique. Update the EDS, oFono and Tracker backends to
use their persona IIDs as display IDs, instead full names. This isn’t an
API break, as the API was always documented as being unique.

https://bugzilla.gnome.org/show_bug.cgi?id=711827