bluez: Correctly handle OBEX transfers which take zero time
If an OBEX transfer takes zero time (say, in an ideal world or, perhaps,
in a unit test) the transfer Status property will start out as
‘complete’ and never change. Previously, completed transfers were only
detected on receipt of a property change notification which, in this
situation, never came. Now, the initial value of the Status property is
also checked.

https://bugzilla.gnome.org/show_bug.cgi?id=712274