eds: Move with-session-bus-eds.sh into test-case.vala
Eliminate some more of the shell scripting which plagues our test cases.
This wipes up to 2s off the startup time for the EDS tests (which was
previously hard-coded to spend waiting for the EDS processes to start up;
now the processes are started through D-Bus service activation, so we only
wait as long as necessary).