bluez: Implement [enable|disable]_persona_store()
Disable all persona stores by default in the BlueZ backend, since they
should only be connected to when explicitly requested by the user. This
can only be programmatically achieved at the moment by calling
bluez_backend.set_persona_stores(["00:11:22:33:44:55"]), where the
string is the Bluetooth address of the phone to enable. (Multiple phones
may be specified.)

This is a partial solution to the problem. A complete solution would
involve creating PersonaStores for all eligible Bluetooth devices, then
only allowing prepare() on the ones which are enabled by the user.
That’s still under discussion in the bug; this is an interim fix to
ensure folks doesn’t continually cause “do you want to allow your
computer access to this phone’s address book” messages to pop up on
people’s phones.

https://bugzilla.gnome.org/show_bug.cgi?id=720274