build: Don’t link backend libraries with libfolks-internal.la helper
Because libfolks-internal.la is a libtool helper library, and is never
actually installed, it gets statically linked into each backend. This
means that the GObject types in it (such as FolksSmallSet) get
duplicated between libfolks.so and each backend. This causes GObject
errors when the backends are loaded and attempt to re-register types
which already exist.

This is a partial revert of commits:
 • 150858a0237291d8269d222defcb82f54c1125cf
 • ef76b7cacb5cfdf7963e5c2a2fe3bc8c1c97838e