tests: Add private copy of GTestDBus
This is due to the suggested deprecation of GTestDBus, and the refusal
of upstream to accept patches which add system bus support. Those
patches are a requirement for folks to use GTestDBus to test its BlueZ
backend.

Ideally, folks could be ported to some other D-Bus mocking framework,
but a lot of work has been invested in the GTestDBus approach, and it’s
currently bit rotting something horrible. I just don’t have time at the
moment to port it all over to something else; so copying the GTestDBus
sources locally and modifying them is the most pragmatic approach.

gtestdbus.[ch] are licensed under LGPLv2.1+ (same as libfolks), and the
copyright is held by:
 • Copyright © 2008–2010 Red Hat, Inc.
 • Copyright © 2012 Collabora Ltd.
Authors:
 • David Zeuthen <davidz@redhat.com>
 • Xavier Claessens <xavier.claessens@collabora.co.uk>