tests: Add support for starting isolated system buses
Add a new FOLKS_TEST_DBUS_SYSTEM_BUS flag which can be passed to
folks_test_dbus_new() to create a system bus instead of a session bus. This
handles setting the correct environment variables, and ensures that the
bus’ security policy is completely permissive.

This is a tweaked version of the patch from
bug: https://bugzilla.gnome.org/show_bug.cgi?id=712148