tests: Add a FOLKS_TEST_DBUS_SESSION_BUS flag
This mirrors the FOLKS_TEST_DBUS_SYSTEM_BUS flag, and can be used to clarify
which bus is expected to be created. The default behaviour (if
FOLKS_TEST_DBUS_SYSTEM_BUS isn’t specified) is to create a session bus,
however, so setting this flag isn’t at all required.

This is a tweaked version of the patch from
bug: https://bugzilla.gnome.org/show_bug.cgi?id=712148