Make warnings non-fatal during "final teardown"
Historically, the Folks tests have not been expected to cope gracefully
with dropping off the session bus. This frequently results in warnings,
which default to being fatal. For instance, enable-disable-stores
fails with this fatal warning:

folks-WARNING **: Error preparing persona store 'eds:test': Couldn't get view for address book ‘test’: The connection is closed

The long-term solution would be to make sure that disconnection from
the bus is an identifiable error domain/code that gets propagated
through the stack, but that would have to be done individually
per-backend.

https://bugzilla.gnome.org/show_bug.cgi?id=726973