core: Fix FOLKS_DISABLE_LINKING
The second step of the _add_personas() function looks at candidate_inds,
which is currently populated even if linking is disabled (in a failed
attempt to keep all the FOLKS_DISABLE_LINKING logic to a single
if-statement at the bottom of the function).

This patch takes a different approach, and ensures that candidate_inds
is not populated if FOLKS_DISABLE_LINKING is enabled, adding an
assertion to try and prevent this bug being re-introduced in future.

A test case will be added in a future commit (see the bug report).

https://bugzilla.gnome.org/show_bug.cgi?id=727405