eds: Bypass test failures on GTestDBus teardown
GTestDBus can sometimes fail on teardown, causing our EDS tests to
spuriously fail, and this has become a real thorn in our side. While we
wait for the proper solution to the relevant GLib bugs to be hashed out,
let’s simply exit early from the tests to ignore failures in teardown.

This will not affect (non-teardown) test failures from true positives,
as it only happens on final teardown.

https://bugzilla.gnome.org/show_bug.cgi?id=726973