bluez: Mark photos as initially up-to-date
The BlueZ backend only downloads contact photos if it determines its set
of photos is out of date (since BlueZ only lets it download _all_
photos, or _no_ photos). However, the variable controlling this could
never actually be set to true, so the backend always thought its set of
photos was outdated, and always re-downloaded them.