build: Disable libsocialweb backend on distcheck
libsocialweb is effectively dead upstream, with no releases or
meaningful code changes in two years. I can no longer build it on a
modern GNOME development system, and it’s unlikely that it works as
described any more due to changes in the web services.

If users or packagers want to re-enable the backend, they will have to
build from source with configure flags:
    --enable-vala --enable-libsocialweb-backend
and with a Vala compiler available.