tests: Port to use installed-tests
Port the unit tests to follow the installed-tests standard, installing
the test binaries and generated .test metadata files.

This removes the existing --enable-tests configure option, replacing it
with the standard:
 • --enable-modular-tests
 • --enable-installed-tests
options. --enable-modular-tests is a direct replacement for
--enable-tests, controlling whether tests are built at compile time, or
only when run under `make check`. --enable-installed-tests controls
whether tests will be installed on the system.

If the tests are installed, use
    gnome-desktop-testing-runner folks
to run them all.

https://bugzilla.gnome.org/show_bug.cgi?id=743398