inspect: Ensure terminal state is restored if PAGER cannot be spawned
If $PAGER cannot be spawned (e.g. due to not existing, or due to the
user not having permission to run it), we would previously leave the
terminal with the readline handler uninstalled. Correctly restore the
readline state instead, and redisplay the current prompt.

This avoids leaving the user with no way to enter text or kill
folks-inspect in cases similar to
    PAGER=really-doesnt-exist folks-inspect