tests: Fix race condition in multi-store EDS tests
If a single EDS test used multiple persona stores (i.e. address books,
each of which is a single ESource), it was possible for creation of the
ESources to race so that the ESourceRegistry notified of creation of the
sources in a different order from the calls to Eds.Backend.set_up().
This caused Eds.Backend.set_up() to return prematurely for some of the
persona stores, causing failures later on.

Check the UID of ESources which have been signalled as created to ensure
the correct one is chosen for each set_up() call.