eds: Pass contact update notifications through idle queue
This is a follow-up to commit e397d869d8f84641825252657aa34ae8487dd526.
If the main context queue was non-empty, the idle callback from
_contacts_changed_cb() would be queued for a while, whereas the callback
in _commit_modified_property() would be executed immediately. This means
_commit_modified_property() would finish yielding before the contact in
question had been updated from EDS (the updated vCard would be sitting
in the idle queue at this point), which is wrong.

Fix that by queueing the callback in _commit_modified_property() in the
idle queue, to keep it in order with the _contacts_changed_cb()
callback.