eds: Fix race condition when adding contacts
Similarly to with property modifications, there is a race condition when
adding contacts to EDS: the EBookClient.add_contact() method returns
once the EDS server has allocated a UID for the contact, but this may be
before folks has received the objects-added signal with the new
contact's updated vCard. This vCard has potentially been modified by the
server (e.g. by adding a VERSION property) from the version initially
submitted by folks; so it is important that folks uses this to create
its Edsf.Persona. It was not doing so previously — there were two code
paths adding the Edsf.Persona to the store, which would race with
slightly different results.

The fix is to factor out the add_contact() call and yield until the new
contact is seen in an objects-added signal emission.