eds: Handle removed contacts in objects-added signal handler
This is a very unlikely situation, but it may be possible for a contact
with the same UID as an existing one in the store to be received from
the server in an objects-added signal. If so, folks' internal state
would previously be corrupted. Fix this by signalling removal of the
previous persona before adding the new one for that UID.