tests: Eliminate non-fatal timeouts from the EDS tests
Consistently use TestUtils.loop_run_with_timeout() instead of
TestUtils.loop_run_with_non_fatal_timeout(). Fix a number of race
conditions where (for example) individuals-changed can be emitted as
part of teardown of the test, while the test’s signal handler is still
connected to it, resulting in an assertion failure because the handler
does not expect personas to be removed.

This should also speed the tests up a bit, because they’re no longer
waiting for a timeout before checking for success. They now quit the
main loop on success.