eds: Eliminate another race condition when committing contacts
We can’t connect to the objects_modified signal, as we can’t guarantee
whether the handler installed in _commit_modified_property() or the
_contacts_changed_cb() handler will be called first. Since they both
enqueue to the idle queue in order, the callback from
_commit_modified_property() will still operate on an outdated
Edsf.Persona if it is called first.

This race condition can only happen when _commit_modified_property() is
called with a null property_name, which happens for custom properties
and extended fields.

Fix this by listening for changes to the Edsf.Persona.contact property
instead of objects_modified. This is always guaranteed to change when a
Persona is updated (because it represents the current vCard), and fixes
the ordering problem, because the Persona’s properties will have been
updated by the time the contact property is notified (see
Edsf.Persona._update() for details).