telepathy: Improve error handling on connection failure of a CM
If a CM is initially offline when the Tpf.PersonaStore is created,
calling this.account.connection.prepare_async() will start trying to
connect it. If the host is unreachable, this will result in waiting for
a network timeout before Telepathy can declare failure.

There are two problems here:
 • This network timeout may be longer than the quiescence timeout built
in to the IndividualAggregator. There’s not much we can do about that.
 • The GError was not previously handled properly, so the PersonaStore
would never reach quiescence anyway.

Fix this by falling back to loading the cache if connecting fails.