bluez: Save/Restore list of enabled BlueZ persona stores
This allows us to persistently enable or disable particular Bluetooth
devices for providing contact data. The idea is that a UI (such as
gnome-contacts) could use [enable|disable]_persona_store() to set the
list of enabled devices, so that even if a device is paired and supports
PBAP, it doesn’t persistently ask for permission to send contacts to
your laptop when you don’t want it to.

https://bugzilla.gnome.org/show_bug.cgi?id=720274