backends: Fix libsocialweb and tracker backends with recent valac
The type checking got better in recent versions of valac, so the
ambiguous types of the personas property in the PersonaStore
implementations in these backends now throw an error. Fix that by making
the type more specific.

This should not be an API break; these classes are still derived from
Folks.PersonaStore, where the types have not changed and were never
ambiguous.