eds: Ignore contacts with no UID set in their vCard
This can happen with contacts from some CardDAV servers (like Radicale).
Since we use the UID as part of the persona’s IID, this can result in
non-unique IIDs within a single persona store, which the rest of the
code in libfolks does not expect. This can result in those personas
being linked together (due to sharing IIDs) incorrectly by the
aggregator.

Instead, ignore any contacts which have no UID set. This is not a
perfect solution (since UIDs are optional in the vCard specification,
which we would ideally handle), but the alternative is to invent an ID
which we then potentially couldn’t propagate back to the server, so
could not guarantee its stability over time.

https://bugzilla.gnome.org/show_bug.cgi?id=773011