core: Remove locking everywhere
folks is single-threaded, and not thread-safe. It never has been. The
locking was added ‘because it won’t hurt’ — but it actually generates
invalid code which double-unlocks mutexes. This doesn’t hurt (in that it
doesn’t crash), but it certainly isn’t valid, and (rightfully) generates
a lot of ThreadSanitizer noise.

https://bugzilla.gnome.org/show_bug.cgi?id=778005