tests: Enable required BlueZ devices during unit tests
Since the BlueZ backend has gained the ability to enable and disable
devices based on their Bluetooth address, and to load that state from a
file, the unit tests were broken because all devices are disabled by
default. Fix the tests by writing out a mock configuration file which
enables the mock devices we want to use.

Signed-off-by: Philip Withnall <withnall@endlessm.com>