Linking: store linking personas in key-file
This stores the linking personas in the key file store instead of the
primary store. The primary store can be anything writable but we
shouldn't add any personas the the users store, escpecally since we
can't safely remove it again. Therefore this makes always use of the
key-file store to save the linking persona and personas needed if a not
writable persona is involved in linking and unlinking.