Prevent some copies by using `var` less
The problem with Vala's `var` keyword is that it always takes a strong
reference, even when that is unnecessary (e.g. because you won't change
the variable, or have reference counting in only one thread). Especially
for certain boxed types and strings this is not really good, as they
might trigger full copies of the data.