Don't use "with" as identifier
The latest Vala compiler adds `with` as a new keyword. That means that
suddenly, all Vala code using "with" as an identifier can suddenly
break. Even funnier, if you expose this in a VAPI (like folks does),
then all the projects that depend on it will also fail to build :-)

Probably this should be guarded more properly by `valac`, but let's be a
bit lenient on our side and just rename the variable.