Cut down on the amount of owned variables
In a lot of places in Folks, we unnecessarily create a lot of owned
variables. Although the performance impact of a string copy or a
refcount increment is normally completely not a concern, we also do this
in a lot of hot code paths (e.g., for each contact that is loaded). This
in turn mostly impacts larger address books.