tracker: Drop Tracker code
The story of saved contacts in Tracker is a convoluted one,
and this code is mostly an artifact. 65 Million years ago, when
Nokia dominated the world, Maemo/meego relied on Tracker and had
specialized contact apps that used it as a storage.

After Elop hit the Earth, saved contacts barely survived the
event, mostly feeding off a small Evolution plugin. But that
was not meant to last long, the evolution plugin went extinct,
taking all traces of saved contacts with it. Just the ontology
remained, a hollow shell that yes, could be able to hold a
contact, would they still exist.

Modern times arrived, and with them the interest on archeology,
thus this fossilized piece of code is found and taken to the
museum where it belongs.

Should Tracker-using contacts ever exist again, the way to
communicate with them would be radically different, let this code
be a testimony of how it won't be done.

Fixes: https://gitlab.gnome.org/GNOME/folks/-/issues/126