tests/bluez: Follow PairDevice() API change
Upstream python-dbusmock changed its API for `PairDevice()` to include a
class parameter. For Python code, this API break was solved by providing
a default value for the argument. Since we rely on the D-Bus interface
however and D-Bus doesn't have a concept of default arguments, this just
broke with an error:

> Error setting up mock BlueZ device:
> GDBus.Error:org.freedesktop.DBus.Error.InvalidArgs: Invalid arguments:
> More items found in D-Bus signature than in Python arguments

Add the same default argument in our Vala interface to work around the
problem.

Related links:
* https://github.com/martinpitt/python-dbusmock/pull/45
* https://github.com/martinpitt/python-dbusmock/issues/123