build: add missing API declaration -include to dummy and EDS backends
In 7a9c8133468c1f8eb7fbb0e68a6a7e061b9452fb, the original
`-include redeclare-internal-api.h` was added for autotools. It is
necessary due to vala internals that require guaranteeing a private
ABI's prototype is available when compiling folks, even though vala
doesn't generate that prototype.

During the port to meson, this `-include` was dropped.

Without it, `-Werror=implicit-function-declaration` will cause folks to
fail to compile. Implicit function declarations are a terrible feature
-- they were part of c89, but c99 removed them without a deprecation
because they were too dangerous to use, notable in a language that
doesn't like to remove features even *with* a deprecation.

For over two decades, compilers have, by default, allowed this invalid C
anyways, on the grounds that it commonly existed, but modern compilers
are finally starting to enforce this. It will become a fatal error by
default in gcc 14, and clang 16.

Folks got this right all the way back in 2013; it disappeared by
accident. Let's restore it as originally intended.