individual: keep strong reference to PersonaStores
When an `IndividualAggregator` is finalized, it finalizes its fields one
by one. As the `_stores` field is declared before the `_individuals`
field, there is a chance it will be finalized before that. As we keep
track inside an `Individual` on how many personas we have from a
specific `PersonaStore`, we should make sure to keep a strong reference
to those, so they don't get finalized before we get a chance to check
those out when the invidual gets finalized too.

This fixes several tests that were recently failing.